requirements.txt
Django==4.2.2
djangorestframework==3.14.0
psycopg2==2.9.7
django-rest-framework
markdown
django-filter
Pillow